Refactor repeated circuit block
