+++ surface_sim/circuit_blocks/surface_code_xzzx_google.py
@@ -4,8 +4,7 @@ https://doi.org/10.1038/s41586-022-05434-1
 https://doi.org/10.48550/arXiv.2207.06431 
 """
 
-from itertools import chain, compress
-from typing import List
+from itertools import chain
 import warnings
 
 from qec_util import Layout
@@ -14,18 +13,11 @@ from stim import Circuit, target_rec
 
 from ..models import Model
 
+# methods to have in this script
+from .util import qubit_coords, log_x, log_z
+from .util import init_qubits_xzzx as init_qubits
 
-def qubit_coords(model: Model, layout: Layout) -> Circuit:
-    """Returns a stim circuit that sets up the coordinates
-    of the qubits.
-    """
-    coord_dict = {q: layout.get_coords([q])[0] for q in layout.get_qubits()}
-    circuit = Circuit()
-
-    for instruction in model.qubit_coords(coord_dict):
-        circuit.append(instruction)
-
-    return circuit
+__all__ = ["qubit_coords", "qec_round_with_log_meas", "log_x", "log_z", "qec_round", "init_qubits"]
 
 
 def qec_round_with_log_meas(
@@ -301,54 +293,5 @@ def qec_round(
 
     for targets in det_targets:
         circuit.append("DETECTOR", targets)
-
-    return circuit
-
-
-def init_qubits(
-    model: Model, layout: Layout, data_init: List[int], rot_basis: bool = False
-) -> Circuit:
-    """
-    Returns stim circuit corresponding to a logical initialization
-    of the given model.
-    By default, the logical measurement is in the Z basis.
-    If rot_basis, the logical measurement is in the X basis.
-    """
-    anc_qubits = layout.get_qubits(role="anc")
-    data_qubits = layout.get_qubits(role="data")
-
-    qubits = set(data_qubits + anc_qubits)
-
-    circuit = Circuit()
-    for instruction in model.reset(qubits):
-        circuit.append(instruction)
-    circuit.append("TICK")
-
-    exc_qubits = set(compress(data_qubits, data_init))
-    if exc_qubits:
-        for instruction in model.x_gate(exc_qubits):
-            circuit.append(instruction)
-
-    idle_qubits = qubits - exc_qubits
-    for instruction in model.idle(idle_qubits):
-        circuit.append(instruction)
-    circuit.append("TICK")
-
-    stab_type = "x_type" if rot_basis else "z_type"
-    stab_qubits = layout.get_qubits(role="anc", stab_type=stab_type)
-
-    rot_qubits = set()
-    for direction in ("north_west", "south_east"):
-        neighbors = layout.get_neighbors(stab_qubits, direction=direction)
-        rot_qubits.update(neighbors)
-
-    for instruction in model.hadamard(rot_qubits):
-        circuit.append(instruction)
-
-    idle_qubits = qubits - rot_qubits
-
-    for instruction in model.idle(idle_qubits):
-        circuit.append(instruction)
-    circuit.append("TICK")
 
     return circuit
